+++ src/features/manageProfileSettings/components/ProfileSettingsEditor/UnusedComponent.tsx
@@ -4,6 +4,7 @@ import { BaseComponent, ComponentType } from './BaseComponent';
 
 type UnusedComponentProps = {
   component: ProfileSettingComponent;
+  moveFn?: () => void;
 };
 
 export const UnusedComponent: FC<UnusedComponentProps> = props => {